refactor(parser) Move CTO parser/printer out of core into its own package
