+++ packages/concerto-core/api.txt
@@ -114,12 +114,6 @@ class MetaModel {
    + object resolveMetaModel(object,object) 
    + object modelFileToMetaModel(object,boolean?) 
    + object modelManagerToMetaModel(object,boolean?,boolean?) 
-   + string decoratorArgFromMetaModel(object) 
-   + string decoratorFromMetaModel(object) 
-   + string decoratorsFromMetaModel(object,string) 
-   + string propertyFromMetaModel(object) 
-   + string declFromMetaModel(object) 
-   + string ctoFromMetaModel(object,boolean?) 
    + object modelManagerFromMetaModel(object,boolean?) 
    + object ctoToMetaModel(string,boolean?) 
    + object ctoToMetaModelAndResolve(modelManager,string,boolean?)
