+++ packages/concerto-parser/.eslintrc.yml
@@ -1,0 +1,47 @@
+env:
+    es6: true
+    node: true
+    mocha: true
+extends: 'eslint:recommended'
+parserOptions:
+    ecmaVersion: 13
+    sourceType: script
+rules:
+    indent:
+        - error
+        - 4
+    linebreak-style:
+        - warn
+        - unix
+    quotes:
+        - error
+        - single
+    semi:
+        - error
+        - always
+    no-unused-vars:
+        - error
+        - args: none
+    no-console: warn
+    curly: error
+    eqeqeq: error
+    no-throw-literal: error
+    strict: error
+    no-var: error
+    dot-notation: error
+    no-tabs: error
+    no-trailing-spaces: error
+    # no-use-before-define: error
+    no-useless-call: error
+    no-with: error
+    operator-linebreak: error
+    require-jsdoc:
+        - error
+        - require:
+            ClassDeclaration: true
+            MethodDefinition: true
+            FunctionDeclaration: true
+    valid-jsdoc:
+        - error
+        - requireReturn: false
+    yoda: error
