+++ packages/concerto-parser/.npmignore
@@ -1,0 +1,90 @@
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# 
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+# Logs
+logs
+*.log
+npm-debug.log*
+
+# Runtime data
+pids
+*.pid
+*.seed
+
+# Directory for instrumented libs generated by jscoverage/JSCover
+lib-cov
+
+# hidden config files
+.editorconfig
+.eslint.yml
+.eslintignore
+.eslintrc.yml
+.tern-project
+
+# test directory
+test
+
+# scripts directory
+scripts
+
+# Coverage directory used by tools like istanbul
+coverage
+
+# nyc test coverage
+.nyc_output
+
+# Grunt intermediate storage (http://gruntjs.com/creating-plugins#storing-task-files)
+.grunt
+
+# node-waf configuration
+.lock-wscript
+
+# Compiled binary addons (http://nodejs.org/api/addons.html)
+build/Release
+
+# Dependency directories
+node_modules
+jspm_packages
+
+# Optional npm cache directory
+.npm
+
+# Optional REPL history
+.node_repl_history
+
+# JSDoc
+out
+
+# Mac files.
+**/.DS_Store
+
+*.swp
+
+# Build generated files should be ignored by git, but not by npm.
+# lib/acl/parser.js
+# lib/introspect/parser.js
+# lib/query/parser.js
+# index.d.ts
+
+.babelrc
+api.txt
+changelog.txt
+jsdoc.json
+HEADER
+HEADER.md
+rollup.config.js
+lib/introspect/parser.pegjs
+
+# Locally built npm packages
+*.tgz
